refactor ping test to work with the keepalive loop, and shorten the time taken to test it
