+++ pkg/p2p/peer/peer_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus/agreement"
 	"github.com/dusk-network/dusk-blockchain/pkg/core/tests/helper"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/peer"
-	"github.com/dusk-network/dusk-blockchain/pkg/p2p/peer/dupemap"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/peer/processing"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/peer/processing/chainsync"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/protocol"
@@ -109,38 +108,6 @@ func TestWriteLoop(t *testing.T) {
 	decoded = decoded[4:]
 
 	assert.Equal(t, decoded, buf.Bytes())
-}
-
-// Test that the 'ping' message is sent correctly, and that a 'pong' message will result.
-func TestPingLoop(t *testing.T) {
-	bus := eventbus.New()
-	client, srv := net.Pipe()
-
-	responseChan := make(chan *bytes.Buffer, 10)
-	writer := peer.NewWriter(client, processing.NewGossip(protocol.TestNet), bus)
-	go writer.Serve(responseChan, make(chan struct{}, 1))
-
-	// Set up the other end of the exchange
-	responseChan2 := make(chan *bytes.Buffer, 10)
-	writer2 := peer.NewWriter(srv, processing.NewGossip(protocol.TestNet), bus)
-	go writer2.Serve(responseChan2, make(chan struct{}, 1))
-	// Give the goroutine some time to start
-	time.Sleep(100 * time.Millisecond)
-
-	reader, err := peer.NewReader(client, processing.NewGossip(protocol.TestNet), dupemap.NewDupeMap(0), bus, rpcbus.New(), &chainsync.Counter{}, responseChan2, make(chan struct{}, 1))
-	if err != nil {
-		t.Fatal(err)
-	}
-	go reader.ReadLoop()
-
-	// We should eventually get a pong message out of responseChan2
-	buf := <-responseChan2
-	topic, err := topics.Extract(buf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	assert.Equal(t, topics.Pong.String(), topic.String())
 }
 
 func BenchmarkWriter(b *testing.B) {
